a.i_project/requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
PyPDF2
groq
python-dotenv
numpy